src/requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.4.1
atomicwrites==1.4.1
attrs==21.4.0
autopep8==1.6.0
blessed==1.19.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.9
colorama==0.4.6
configparser==5.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.10
django-cleanup==7.0.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-oauth-toolkit==1.6.1
django-picklefield==3.0.1
django-q==1.3.9
django-q-email==5.0.1
django-seed==0.3.1
django-webpack-loader==1.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-jwt==1.19.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==10.0.0
filelock==3.4.1
google-auth==2.3.3
identify==2.4.4
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.2.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
jinxed==1.2.0
jwcrypto==1.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.0
nodeenv==1.6.0
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.1.5
phonenumbers==8.12.42
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.17.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.14.0
PyJWT==1.7.1
pyparsing==3.0.6
pytest==7.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
reportlab==3.6.6
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
toposort==1.7
twilio==6.60.0
typing-extensions==4.0.1
Unidecode==1.3.3
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.13.1
wcwidth==0.2.5
wrapt==1.13.3
xlrd==2.0.1
zipp==3.6.0
